Move HLS binary commands into moq-cli
