+++ rs/moq-srt/README.md
@@ -7,7 +7,7 @@ by its stream id `m=` mode:
 
 - `m=publish` (the default): ingest. Demux the connection's transport stream
   with [`moq-mux`](../moq-mux) and publish it into a MoQ origin as an ordinary
-  broadcast. The contribution-ingest analogue of `moq-hls`'s import and
+  broadcast. The contribution-ingest analogue of `moq-cli hls import` and
   `moq-rtc`'s WHIP.
 - `m=request`: egress. Re-mux a broadcast from the origin back to MPEG-TS and
   stream it to the caller, so `vlc srt://...` and `ffmpeg -i srt://...` can play
@@ -54,7 +54,7 @@ moq-srt serve --server-bind [::]:443 --tls-generate localhost \
 ```
 
 `publish` instead forwards every ingested broadcast out to a remote relay over
-WebTransport (like `moq-hls import`):
+WebTransport (like `moq-cli hls import`):
 
 ```bash
 moq-srt publish --relay https://relay.example.com \
